Update 04.16(화요일 새벽) - Sound 작업중 FMOD를 이용한 GameSound 구현중 SoundPlayer 컴포넌트 클래스 추가 Sound 클래스 추가 SoundPlayer 컴포넌트에서 MySoundPlayer() 함수를 통해 사운드를 플레이 시킬수있게

+++ Project_Alice/Project/Alice_Engine/01.Core/Input.cpp
@@ -230,8 +230,8 @@ void CInput::ComputeMouse(bool bStart)
 	RECT	rcClient = {};
 	GetClientRect(WINDOWHANDLE, &rcClient);
 
-	float	x = rcClient.right - rcClient.left;
-	float	y = rcClient.bottom - rcClient.top;
+	float	x = (float)(rcClient.right - rcClient.left);
+	float	y = (float)(rcClient.bottom - rcClient.top);
 
 	RESOLUTION	tRS = _RESOLUTION;
 
@@ -253,5 +253,5 @@ void CInput::ComputeMouse(bool bStart)
 
 	m_tMousePos = tMousePos;
 
-	m_pMouseTr->SetWorldPos(tMousePos.x, tMousePos.y, 0.f);
+	m_pMouseTr->SetWorldPos((float)tMousePos.x, (float)tMousePos.y, 0.f);
 }
